Update gem versions
Updating the versions of many gems -- specifically, ffi and mysql2. But allowing the bundle tool to update anything else it wanted to. This was necessary for my local install, because the explicit version of ffi set by stash_discovery was conflicting with the version set in Gemfile.lock here, and the (older) version of mysql2 was no longer compiling.

@dmoles suggested that this may not be necessary if we run
`./travis-build.sh —bundle-only` from the `stash` root, but I wasn't able to get that approach working, so I'm putting this change up for discussion.