Modified spec_helper.rb so it loads all the items that Rails normally does.  It seems like the tests/spec files are not loading as easily as they should.
It's true that engines are harder to test with the test apps and similar, but this kind of difficulty makes me thing something isn't right.

David started doing these hacks and I'm following in his footsteps so as not to start from scratch.  But ouch.

It would be good to clean up the tests and figure out if we're not doing something the normal/right way since I can't image it should really be so hard for normal projects.