This is a case of both a sharing link and also a large package download at the same time.
A fix for a very simple code change, I'll see if I can add tests for this (unit or request?) since the UI is tricky to test.

This was a simple, tiny change in our code on production, so made the change manually in the code to keep sharing downloads from failing for users and then restarted passenger.